requirements.txt
numpy~=1.20.3
torch~=1.10.0
PyYAML~=5.4
scilpy~=1.2.2
dipy~=1.3.0
h5py~=2.10.0
nibabel~=3.0.2